refactor(repo C3): extract store init, maintenance, and handelization
Split DB lifecycle into three files: store-init.ts (initializeDatabase,
ensureVectorIndexIntegrity, isSqliteVecAvailable, verifySqliteVecLoaded,
ensureVecTableInternal, and the _sqliteVecAvailable module flag),
store-maintenance.ts (vacuum, WAL checkpoint, sidecar cleanup, index health,
capability lookup), and handelize.ts (getDocid, handelize, emojiToHex).
createStore + Store type stay in engine/repository.ts because they wire
together symbols from every other cluster.

Build: clean. Tests: 970/970 pass.
